--- stock_analysis/tests.py
@@ -1,48 +1,0 @@
-"""Tests for stock analysis application. All function and unit tests."""
-
-
-import pytest
-
-"""
-Function/Class list:
-
-stock_analysis/models/__init__.py :
-
-    get_engine()
-    get_session_factory()
-    get_tm_session()
-    includeme()
-
-stock_analysis/models/mymodel.py :
-
-    User()
-    Portfolio()
-
-stock_analysis/scripts/initialzedb.py :
-
-    usage()
-    main()
-
-stock_analysis/views/default.py :
-
-    home_view()
-    detail_view()
-    portfolio_view()
-    get_symbol()
-    logout()
-    process_symbol() ?
-
-stock_analysis/__init__.py :
-
-    main()
-
-stock_analysis/routes.py :
-
-    includeme()
-
-stock_analysis/security.py :
-
-    includeme()
-    SecRoot()
-    is_authorized()
-"""
